data_generator/requirements.txt
numpy-stl==2.16.3
pymadcad==0.11.0
pandas==1.4.3
